src/main/java/tan/fandbaispring/ai-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-google-genai>=2.1.12
langchain-openai>=0.0.2
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
